feat: 상품 삭제하기 API 연동

+++ src/components/modal/DeleteModal.tsx
@@ -2,40 +2,56 @@ import { UseInfiniteQueryResult } from "@tanstack/react-query";
 import ModalContainer from "../app/ModalContainer";
 import styles from "@/styles/app/modal.module.css";
 import useDelProductComments from "@/src/hooks/useDelProductComment";
+import useDelProduct from "@/src/hooks/useDelProduct";
+import { useRouter } from "next/router";
 
 interface Props {
   productId?: string | string[];
   commentId?: number;
+  isDelProduct?: boolean;
   isDelProductComment?: boolean;
   showModal: boolean;
   setShowModal: (value: boolean) => void;
   isModalMessage: string;
-  refetch: UseInfiniteQueryResult["refetch"];
+  refetch?: UseInfiniteQueryResult["refetch"];
 }
 
-export const DelCommentModal = ({
+export const DeleteModal = ({
   productId,
   commentId,
+  isDelProduct,
   isDelProductComment,
   showModal,
   setShowModal,
   isModalMessage,
   refetch,
 }: Props) => {
+  const router = useRouter();
+
+  const { mutate: deleteProduct } = useDelProduct(productId);
   const { mutate: deleteComment } = useDelProductComments(productId!);
 
   const handleConfirmModal = () => {
-    if (!commentId || !productId) return;
+    if (isDelProduct) {
+      if (!productId) return;
 
-    deleteComment(
-      { commentId },
-      {
-        onSuccess: async () => {
-          await refetch();
-          setShowModal(false);
-        },
-      }
-    );
+      deleteProduct();
+      router.replace("/items");
+    }
+
+    if (isDelProductComment) {
+      if (!commentId || !productId) return;
+
+      deleteComment(
+        { commentId },
+        {
+          onSuccess: async () => {
+            if (refetch) await refetch();
+            setShowModal(false);
+          },
+        }
+      );
+    }
   };
 
   return (
@@ -44,14 +60,12 @@ export const DelCommentModal = ({
         <div className={styles.contents}>
           <div className={styles.message}>{isModalMessage}</div>
           <div className={styles.buttonContainer}>
-            {isDelProductComment && (
-              <button
-                className={styles.cancel}
-                onClick={() => setShowModal(false)}
-              >
-                취소
-              </button>
-            )}
+            <button
+              className={styles.cancel}
+              onClick={() => setShowModal(false)}
+            >
+              취소
+            </button>
             <button type="submit" onClick={handleConfirmModal} autoFocus>
               확인
             </button>
